Loki: Rename public/app/plugins/datasource/loki/live_streams_result_transformer.ts file to follow convention
