+++ public/app/plugins/datasource/loki/liveStreamsResultTransformer.test.ts
@@ -1,6 +1,6 @@
 import { CircularDataFrame, FieldType } from '@grafana/data';
 
-import * as ResultTransformer from './live_streams_result_transformer';
+import * as ResultTransformer from './liveStreamsResultTransformer';
 import { LokiTailResponse } from './types';
 
 describe('loki result transformer', () => {
