+++ public/app/plugins/datasource/loki/liveStreamsResultTransformer.ts
