+++ public/app/plugins/datasource/loki/live_streams.ts
@@ -4,7 +4,7 @@ import { webSocket } from 'rxjs/webSocket';
 
 import { DataFrame, FieldType, KeyValue, CircularDataFrame } from '@grafana/data';
 
-import { appendResponseToBufferedData } from './live_streams_result_transformer';
+import { appendResponseToBufferedData } from './liveStreamsResultTransformer';
 import { LokiTailResponse } from './types';
 
 /**
